requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
google-generativeai==0.3.2
pinecone-client==2.2.4
langchain==0.1.0
langchain-google-genai==0.0.5
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
email-validator==2.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3 